Refactored edu.uwb.braingrid.workbench.data.project to edu.uwb.braingrid.workbench.project
This commit will help to move all components in the workbench toward an MVC structure. While the code has been moving there already, the packages should reflect this change.
